[Refactor] Refactor Comment, AddShare, MyMgmtCard, UpdatePopup, carousel
